dsl/requirements.txt
websocket-client==1.6.4
python-dotenv==1.0.0 